Rename the packages as mom6_forge
