+++ .docstr/examples.rst
@@ -1,22 +1,22 @@
 Examples
 ======================================
 
-*Check out the following Jupyter Notebooks on mom6_bathy GitHub repository:*
+*Check out the following Jupyter Notebooks on mom6_forge GitHub repository:*
 
 1. Spherical grid with an idealized bathymetry:
-`1_spherical_grid.ipynb <https://github.com/NCAR/mom6_bathy/blob/master/notebooks/1_spherical_grid.ipynb>`_
+`1_spherical_grid.ipynb <https://github.com/NCAR/mom6_forge/blob/master/notebooks/1_spherical_grid.ipynb>`_
 
 2. Equatorial refinement:
-`2_equatorial_res.ipynb <https://github.com/NCAR/mom6_bathy/blob/master/notebooks/2_equatorial_res.ipynb>`_
+`2_equatorial_res.ipynb <https://github.com/NCAR/mom6_forge/blob/master/notebooks/2_equatorial_res.ipynb>`_
 
 3. Custom bathymetry and ridges:
-`3_custom_bathy.ipynb <https://github.com/NCAR/mom6_bathy/blob/master/notebooks/3_custom_bathy.ipynb>`_
+`3_custom_bathy.ipynb <https://github.com/NCAR/mom6_forge/blob/master/notebooks/3_custom_bathy.ipynb>`_
 
 4. Apply existing land mask:
-`4_ingest_landmask.ipynb <https://github.com/NCAR/mom6_bathy/blob/master/notebooks/4_ingest_landmask.ipynb>`_
+`4_ingest_landmask.ipynb <https://github.com/NCAR/mom6_forge/blob/master/notebooks/4_ingest_landmask.ipynb>`_
 
 5. Modify existing MOM6 bathymetry via point-and-click widget:
-`5_modify_existing.ipynb <https://github.com/NCAR/mom6_bathy/blob/master/notebooks/5_modify_existing.ipynb>`_
+`5_modify_existing.ipynb <https://github.com/NCAR/mom6_forge/blob/master/notebooks/5_modify_existing.ipynb>`_
 
 6. Use domain configurators:
-`6_demo_editors.ipynb <https://github.com/NCAR/mom6_bathy/blob/master/notebooks/6_demo_editors.ipynb>`_
+`6_demo_editors.ipynb <https://github.com/NCAR/mom6_forge/blob/master/notebooks/6_demo_editors.ipynb>`_
