+++ .docstr/installation.rst
@@ -11,32 +11,32 @@ https://docs.conda.io/en/latest/miniconda.html
 Instructions
 -------------
 
-First, clone the `mom6_bathy` GitHub repository as follows:
+First, clone the `mom6_forge` GitHub repository as follows:
 
 .. code-block:: bash
 
-    git clone --recursive https://github.com/NCAR/mom6_bathy.git
+    git clone --recursive https://github.com/NCAR/mom6_forge.git
 
-Then, `cd` into your newly checked out `mom6_bathy` clone and run the
-following command to install `mom6_bathy` and all dependencies.
+Then, `cd` into your newly checked out `mom6_forge` clone and run the
+following command to install `mom6_forge` and all dependencies.
 
 .. code-block:: bash
 
-    cd mom6_bathy
+    cd mom6_forge
     conda env create -f environment.yml
 
-The above command creates a new conda environment called `mom6_bathy_env`. You can
+The above command creates a new conda environment called `mom6_forge`. You can
 activate this environment by running:
 
 .. code-block:: bash
 
-    conda activate mom6_bathy_env
+    conda activate mom6_forge
 
 To confirm that the installation was successful, execute the following command:
 
 .. code-block:: bash
 
-    python -c "import mom6_bathy"
+    python -c "import mom6_forge"
 
 If no error message is displayed, then the installation is successful. Note that
-you will need to activate the `mom6_bathy_env` environment before every use.
+you will need to activate the `mom6_forge` environment before every use.
