+++ .docstr/mom6_bathy.grid.rst
@@ -1,7 +1,7 @@
 mom6\_bathy.grid module
 ===========================
 
-.. automodule:: mom6_bathy.grid
+.. automodule:: mom6_forge.grid
    :members:
    :undoc-members:
    :show-inheritance:
