+++ .docstr/mom6_bathy.rst
@@ -1,7 +1,7 @@
 mom6\_bathy package
 ===================
 
-.. automodule:: mom6_bathy
+.. automodule:: mom6_forge
    :members:
    :undoc-members:
    :show-inheritance:
@@ -12,5 +12,5 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
-   mom6_bathy.grid
-   mom6_bathy.topo
+   mom6_forge.grid
+   mom6_forge.topo
