+++ .docstr/mom6_bathy.topo.rst
@@ -1,7 +1,7 @@
 mom6\_bathy.topo module
 ============================
 
-.. automodule:: mom6_bathy.topo
+.. automodule:: mom6_forge.topo
    :members:
    :undoc-members:
    :show-inheritance:
