+++ .github/workflows/build-test-docs.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Run pytest
         shell: bash -l {0}
         run: |
-          pytest -v --maxfail=1 --disable-warnings --cov=mom6_bathy --cov-branch --cov-report=xml
+          pytest -v --maxfail=1 --disable-warnings --cov=mom6_forge --cov-branch --cov-report=xml
       - name: Upload coverage report to CodeCov
         uses: codecov/codecov-action@v5
         with:
