+++ mom6_forge/mapping.py
@@ -14,7 +14,7 @@ import xesmf as xe
 MPI = None
 rank = lambda: MPI.COMM_WORLD.Get_rank() if MPI else 0
 
-from mom6_bathy.utils import (
+from mom6_forge.utils import (
     get_mesh_dimensions,
     cell_area_rad,
     normalize_deg,
@@ -508,10 +508,10 @@ def write_mapping_file(
         }
     )
 
-    ds.attrs["title"] = "CESM mapping file generated by mom6_bathy"
+    ds.attrs["title"] = "CESM mapping file generated by mom6_forge"
     # ds.attrs['normalization'] = 'conservative' # todo
     # ds.attrs['map_method'] = 'nearest neighbor smoothing'
-    ds.attrs["history"] = "Generated by mom6_bathy"
+    ds.attrs["history"] = "Generated by mom6_forge"
     ds.attrs["conventions"] = "NCAR-CCSM"
     ds.attrs["date_created"] = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
     if src_mesh_path := src_mesh.encoding.get("source"):
