+++ setup.py
@@ -4,15 +4,15 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="mom6_bathy",  # Replace with your own username
+    name="mom6_forge",  # Replace with your own username
     version="0.0.1",
     author="Alper Altuntas",
     author_email="altuntas@ucar.edu",
     description="MOM6 simple grid and bathymetry generator",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/NCAR/mom6-bathy",
-    packages=["mom6_bathy"],
+    url="https://github.com/NCAR/mom6_forge",
+    packages=["mom6_forge"],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
