+++ tests/test_chl.py
@@ -1,6 +1,6 @@
-from mom6_bathy.grid import Grid
-from mom6_bathy.topo import Topo
-from mom6_bathy.chl import interpolate_and_fill_seawifs
+from mom6_forge.grid import Grid
+from mom6_forge.topo import Topo
+from mom6_forge.chl import interpolate_and_fill_seawifs
 import pytest
 import os
 from utils import on_cisl_machine
